requirements.txt
# Flask
flask>=3.0.0

# Environment
python-dotenv>=1.0.0

# Optional: uncomment for database support
# flask-sqlalchemy>=3.1.0
# flask-migrate>=4.0.0
